+++ libs/common-hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [3.5.1](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.5.0...common-hooks-v3.5.1) (2026-07-08)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.8.0
+  - @cowprotocol/common-const@3.6.0
+
 ## [3.5.0](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.4.1...common-hooks-v3.5.0) (2026-06-25)
 
 ### Minor Changes