+++ libs/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [3.6.0](https://github.com/cowprotocol/cowswap/compare/core-v3.5.1...core-v3.6.0) (2026-07-08)
+
+### Minor Changes
+
+- feat: handle rate limiting in safe api calls (#7729)
+
+### Patch Changes
+
+- fix: add reverve protocol token list (#7824)
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.8.0
+  - @cowprotocol/common-const@3.6.0
+
 ## [3.5.1](https://github.com/cowprotocol/cowswap/compare/core-v3.5.0...core-v3.5.1) (2026-06-23)
 
 ### Patch Changes